requirements.txt
graphviz==0.19.2
matplotlib==3.6.3
PyGithub==1.55
pypandoc==1.10
python-dotenv==0.17.0
PyYAML==6.0
requests==2.28.1

